+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 FEATURES:
 * Add support for key_usage to `vault_pki_secret_backend_root_sign_intermediate` ([#2421])(https://github.com/hashicorp/terraform-provider-vault/pull/2421)
+* Add new ephemeral resource `vault_transit_decrypt`
 
 * Add `private_key_wo` and `private_key_wo_version` fields to Snowflake DB secrets engine config ([#2508])(https://github.com/hashicorp/terraform-provider-vault/pull/2508)